Add commented code for setting alpha and relVelMag
These will not work uncommented as-is since `actuatorLineElement` does not
calculate geometric angle of attack and relative velocity until after the first
time step. A way around this could be to either calculate these upon
element contruction, or call `actuatorLineSource::calcEndEffects` at the end
of `addSup`. The latter will not apply end effects to the first time step,
which seems a little hacky.