Added compatibility with multiphase solvers
For multiphase simulations, the user must specify two new coefficients: 'multiPhase' (bool, true means multi-phase solver is used) and 'phaseName' (word, only needed if multiPhase is true, should correspond to the main phase name - usually air- as defined in the 'transportProperties' dictionary). This is done to pick the correct value for kinematic viscosity.